feat(tagging): refactor link icon and add tag icon on search dropdown
