+++ src/client/user/components/CreateUrlModal/CreateLinkForm.tsx
@@ -125,6 +125,7 @@ const CreateLinkForm: FunctionComponent<CreateLinkFormProps> = ({
                     ? theme.palette.primary.dark
                     : theme.palette.background.default
                 }
+                size={16}
               />
               <Typography
                 variant="body2"
