+++ src/client/user/widgets/LinkIcon.tsx
@@ -1,26 +1,36 @@
-import React from 'react'
+import React, { FunctionComponent } from 'react'
 
 type LinkIconProps = {
-  color: string
+  size: number
+  color?: string
 }
 
-export default ({ color = '#384A51' }: LinkIconProps) => {
+const LinkIcon: FunctionComponent<LinkIconProps> = ({
+  size,
+  color = '#384A51',
+}: LinkIconProps) => {
   return (
     <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width={16}
-      height={16}
+      width="18"
+      height="18"
+      viewBox="0 0 18 18"
       fill="none"
-      viewBox="0 0 16 16"
+      xmlns="http://www.w3.org/2000/svg"
+      style={{
+        width: `${size}px`,
+        height: `${size}px`,
+      }}
     >
       <path
+        d="M3.167 14.834c.73.73 1.69 1.096 2.65 1.096.962 0 1.922-.365 2.653-1.096l2.12-2.122-1.06-1.06-2.12 2.121c-.878.875-2.305.877-3.183 0-.877-.877-.877-2.305 0-3.182L6.349 8.47l-1.06-1.06-2.122 2.12c-1.462 1.462-1.462 3.842 0 5.304zM14.834 8.47c1.46-1.462 1.46-3.842 0-5.303-1.463-1.463-3.843-1.461-5.304 0L7.41 5.288l1.06 1.06 2.12-2.12c.878-.876 2.306-.878 3.183 0 .877.877.877 2.304 0 3.181L11.651 9.53l1.06 1.06 2.123-2.12z"
         fill={color}
-        d="M2.815 13.185c.65.65 1.503.975 2.356.975.854 0 1.708-.325 2.358-.975L9.414 11.3l-.943-.942-1.885 1.886c-.78.778-2.048.779-2.829 0-.78-.78-.78-2.05 0-2.83L5.643 7.53l-.942-.943L2.815 8.47c-1.3 1.3-1.3 3.415 0 4.714zm10.37-5.656c1.299-1.3 1.299-3.415 0-4.714-1.3-1.3-3.415-1.3-4.714 0L6.586 4.7l.943.942 1.885-1.886c.78-.778 2.049-.78 2.829 0 .779.78.779 2.049 0 2.829L10.357 8.47l.942.943 1.886-1.885z"
       />
       <path
+        d="M6.348 12.713l-1.06-1.06 6.364-6.365 1.06 1.06-6.364 6.365z"
         fill={color}
-        d="M5.643 11.3l-.943-.943L10.357 4.7l.943.943L5.643 11.3z"
       />
     </svg>
   )
 }
+
+export default LinkIcon
